requirements.txt
psycopg2-binary
SQLAlchemy
requests
discord.py
